Change module path to clevergo.tech/clevergo
