+++ README.md
@@ -1,16 +1,19 @@
 # CleverGo
 [![Build Status](https://travis-ci.org/clevergo/clevergo.svg?branch=master)](https://travis-ci.org/clevergo/clevergo)
 [![Go Report Card](https://goreportcard.com/badge/github.com/clevergo/clevergo)](https://goreportcard.com/report/github.com/clevergo/clevergo)
-[![Go.Dev reference](https://img.shields.io/badge/go.dev-reference-blue?logo=go&logoColor=white)](https://pkg.go.dev/github.com/clevergo/clevergo?tab=doc)
+[![Go.Dev reference](https://img.shields.io/badge/go.dev-reference-blue?logo=go&logoColor=white)](https://pkg.go.dev/clevergo.tech/clevergo?tab=doc)
 [![Release](https://img.shields.io/github/release/clevergo/clevergo.svg?style=flat-square)](https://github.com/clevergo/clevergo/releases)
 [![Sourcegraph](https://sourcegraph.com/github.com/clevergo/clevergo/-/badge.svg)](https://sourcegraph.com/github.com/clevergo/clevergo?badge)
-[![English](https://img.shields.io/badge/doc-English-blue)](https://clevergo.tech/docs/)
-[![简体中文](https://img.shields.io/badge/文档-简体中文-blue)](https://clevergo.tech/zh/docs/)
-[![繁體中文](https://img.shields.io/badge/文檔-繁體中文-blue)](https://clevergo.tech/zh-hant/docs/)
 
 CleverGo is a lightweight, feature rich and trie based high performance HTTP request router.
 
 ![Benchmark](https://clevergo.tech/img/benchmark.png)
+
+## Documentations
+
+- [English](https://clevergo.tech/docs/)
+- [简体中文](https://clevergo.tech/zh/docs/)
+- [繁體中文](https://clevergo.tech/zh-hant/docs/)
 
 ## Features
